requirements.txt
flask
tensorflow-cpu
gunicorn
numpy
keras
pillow
